# backend/docs/swagger.yaml
definitions:
...
  handlers.ForgotPasswordInput:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  handlers.LoginInput:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  handlers.RegisterInput:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  handlers.ResetPasswordInput:
    properties:
      email:
        type: string
      new_password:
        minLength: 6
        type: string
      otp:
        type: string
    required:
    - email
    - new_password
    - otp
    type: object
...
info:
  contact: {}
paths:
  /api/auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Request OTP for admin password reset
      parameters:
      - description: Forgot password info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ForgotPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Forgot password (admin)
      tags:
      - auth
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate admin and return JWT token
      parameters:
      - description: Admin login info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Login admin
      tags:
      - auth
  /api/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new admin user
      parameters:
      - description: Admin registration info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Register admin
      tags:
      - auth
  /api/auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset admin password using OTP
      parameters:
      - description: Reset password info
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ResetPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Reset password (admin)
      tags:
      - auth
...
swagger: "2.0"